debugging, animation and card fronts
debugged some problems with a conflict with classes and id's. Still need to look at the white border around the card and why thats there. animation of the card backs and fronts is now working, i want to add a wobble effect on the card fronts rather than the zoom i think. next steps, better card front images and the animation of them setup better then on the javascript.