changes to the javascript
the javascript i had in place seemed messy and long so have started to re-write it and setup some of the game differently in how the cards will flip and change there css classes to be the face up side of the card.
countdown timer is working, as is the overlay at the start.