update to theme and javascript functions
Taken out the bootstrap theme as the loading order in the dom was making it really hard to get a shuffle function working that didnt put the cards all over the page. added in a function to restart the game when the timer is up.
Got a shuffle function to work so the card face location in now random with every new game that is played.